requirements.txt
numpy
pandas
matplotlib
scikit-learn
opencv-python
pytesseract
tesseract
pillow